+++ mocks/kafka-message-handler-tx.go
@@ -94,8 +94,6 @@ func (fake *KafkaMessageHandlerTx) ConsumeMessageReturnsOnCall(i int, result1 er
 func (fake *KafkaMessageHandlerTx) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
-	fake.consumeMessageMutex.RLock()
-	defer fake.consumeMessageMutex.RUnlock()
 	copiedInvocations := map[string][][]interface{}{}
 	for key, value := range fake.invocations {
 		copiedInvocations[key] = value