+++ mocks/kafka-offset-store.go
@@ -175,10 +175,6 @@ func (fake *KafkaOffsetStore) SetReturnsOnCall(i int, result1 error) {
 func (fake *KafkaOffsetStore) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
-	fake.getMutex.RLock()
-	defer fake.getMutex.RUnlock()
-	fake.setMutex.RLock()
-	defer fake.setMutex.RUnlock()
 	copiedInvocations := map[string][][]interface{}{}
 	for key, value := range fake.invocations {
 		copiedInvocations[key] = value